requirements.txt
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
